miniscript: correct thresh matching
It required trial-and-error based on the try: except: pass patterrn in
order to match correctly, it would otherwise not detect an entire thresh
pattern. That's pretty ugly, and cleaning up the implem also resulted in
a 2x speedup in the test cases.

Also, a thresh can be matched from 4 elements!