miniscript: implement a satisfier
It's not super nicely architectured, and the satisfaction algorithm
isn't correctly nor entirely implemented (we don't consider non-canon
dissats at all so we might miss some sats, we don't have a DONTUSE marker
so we can't discard satisfactions where 2 sub sats both have no sig).
But i think it's decently readable and does what i need to test
revaultd, which are my current two goals :).

Note also that we remove support for hashes inside pk_h() in this commit
in order to keep dissatisfy() not needing a `sat_material` to be passed.