miniscript: fix the malleability properties mess
The malleability attributes in Property were written before i
implemented the actual logic, which opted into a clearer way of using
well-defined variables for tracking them. So remove this leftover.

In addition, there were some leftover of a previous implementation of
malleability analysis that used those properties in the fragments, which
i hadn't properly cleaned up. Do it now.